Add files via upload
Worked on getting the buttons declared and into groups. Still need to display them all on the layout we've got, but the underlying code is there. Wanted to post what I had before bed. Game still works, just didn't want this in with the mix since it's still Game.java